moved IdeasToWebApps comp to ./sections/DataToDesign/ and cleaned file
